Switch switch and clear
