+++ apps/wallet-mobile/src/features/Swap/useCases/StartSwapScreen/CreateOrder/ShowTokenActions/SwitchTokens.tsx
@@ -1,22 +1,15 @@
 import {useSwap} from '@yoroi/swap'
 import React from 'react'
-import {StyleSheet, Text} from 'react-native'
 import {TouchableOpacity} from 'react-native-gesture-handler'
 
-import {COLORS} from '../../../../../../theme'
-import {useStrings} from '../../../../common/strings'
+import {Icon} from '../../../../../../components/Icon'
 
 export const SwitchTokens = () => {
-  const strings = useStrings()
-  const {resetQuantities} = useSwap()
+  const {switchTokens} = useSwap()
 
   return (
-    <TouchableOpacity onPress={resetQuantities}>
-      <Text style={styles.text}>{strings.clear}</Text>
+    <TouchableOpacity onPress={switchTokens}>
+      <Icon.Switch size={24} />
     </TouchableOpacity>
   )
 }
-
-const styles = StyleSheet.create({
-  text: {color: COLORS.SHELLEY_BLUE, fontWeight: '500'},
-})
